登录后可编辑自己的内容，不登录只显示所有的内容

+++ routes/api.js
@@ -5,7 +5,13 @@ var Note = require('../model/note').Note;
 /* GET users listing. */
 router.get('/notes', function(req, res, next) {
   console.log('/notes...')
-  Note.findAll({raw:true}).then((notes)=>{
+  var options = {raw: true}
+  if(req.session.user){
+    options.where = {
+      uid: req.session.user.id
+    }
+  }
+  Note.findAll(options).then((notes)=>{
     console.log(notes)
     res.send({status: 0, data: notes})
   })
@@ -15,8 +21,10 @@ router.post('/notes/add',function(req, res, next){
     return res.send({status: 1,errorMsg: '请先登录'})
   }
 
+  var uid = req.session.user.id
   var note = req.body.note
-  Note.create({text:note}).then(()=>{
+
+  Note.create({text:note, uid: uid}).then(()=>{
     res.send({status:0})
   }).catch(()=>{
     res.send({status: 1, errorMsg: '添加出错'})
@@ -27,7 +35,8 @@ router.post('/notes/edit',function(req, res, next){
   if(!req.session.user){
     return res.send({status: 1,errorMsg: '请先登录'})
   }
-  Note.update({text: req.body.note}, {where: {id: req.body.id}} ).then(()=>{
+  var uid = req.session.user.id
+  Note.update({text: req.body.note}, {where: {id: req.body.id, uid: uid}} ).then(()=>{
     console.log(arguments)
     res.send({status: 0})
   }).catch(()=>{
@@ -39,7 +48,8 @@ router.post('/notes/delete',function(req, res, next){
   if(!req.session.user){
     return res.send({status: 1,errorMsg: '请先登录'})
   }
-  Note.destroy({where:{id:req.body.id}}).then(()=>{
+  var uid = req.session.user.id
+  Note.destroy({where:{id:req.body.id, uid: uid}}).then(()=>{
     res.send({status: 0})
     console.log('false')
   }).catch(()=>{
